% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onthefly-package.R
\docType{package}
\name{onthefly-package}
\alias{onthefly}
\alias{onthefly-package}
\title{onthefly: Simulation-On-The-Fly Likelihood-Free Inference}
\description{
R package to help with creating simulation-on-the-fly based 
    likelihood-free inference engines. Based on ideas in 
    <https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7687905/>.
}
\author{
\strong{Maintainer}: Russell Dinnage \email{r.dinnage@gmail.com} (\href{https://orcid.org/0000-0003-0846-2819}{ORCID}) [copyright holder]

}
\keyword{internal}
