% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{otf_create_trainer}
\alias{otf_create_trainer}
\title{Create an \code{otf_trainer} object for training an on-the-fly likelihood-free
inference engine}
\usage{
otf_create_trainer(sim_fun, train_fun, n_latent, n_accum)
}
\arguments{
\item{sim_fun}{Function that takes a list of parameters and returns a
simulation output}

\item{train_fun}{A function that takes a simulation output as input and
outputs a list of two vectors specifying a stochastic 'latent' code. The
first vector is means of Gaussian distributions, the second is corresponding
standard deviations for the same Gaussian distributions. The function should
also have a second argument \code{n_latent} specifying the number of dimensions
for the latent code. The function generally is a \code{torch} neural network,
usually consisting of a \code{torch} \code{nn_module}.}

\item{n_latent}{The number of latent dimensions for the latent code to use.}

\item{n_accum}{The number of simulations to 'accumulate' in a batch before
running the train_fun}
}
\value{
An \code{otf_trainer} object which can be used to run likelihood-free
inference (using \code{\link[=otf_train]{otf_train()}})
}
\description{
Create an \code{otf_trainer} object for training an on-the-fly likelihood-free
inference engine
}
